Update load.sh
following the example setups from the  book you have a self signed certificate, so curl will complain :
curl failed to verify the legitimacy of the server and therefore could not
establish a secure connection to it. To learn more about this situation and
how to fix it, please visit the web page mentioned above.

Adding the -k option will accept the self signed certificate in TLS